Broke code into separate files
Also update the layout of the testcases to be clearer.
